+++ Common/src/main/java/fuzs/puzzleslib/core/Services.java
@@ -1,9 +1,5 @@
 package fuzs.puzzleslib.core;
 
-import fuzs.puzzleslib.registry.FuelManager;
-
-import java.util.ServiceLoader;
-
 /**
  * common services
  * @deprecated renamed to {@link CoreServices}